requirements.txt
streamlit
pandas
numpy
scikit-learn
openai
python-dotenv
python-docx
PyMuPDF
Pillow

-e .